Fix bug when responding to empty SETUP on EP0
This fixes a bug that's exposed when plugging in the device in a
Windows 10 host. After enumeration it sends a SET CONTROL LINE STATE
request followed by a SET LINE CODING request. The former has 0 bytes
payload data, which caused the USB device to stall the latter
transaction. This happened because USBCS0_DATA_END was set in
usb_ep0_setup (indicating no data expected for this transaction), and
then again in usb_ep0_flush (indicating no data expected in the _next_
transaction). The bit remained set, causing the MCU to stall the
following OUT transaction.

This didn't cause problems on Linux because it has a different access
pattern.

The fix is to call usb_ep0_flush() only when needed, i.e. when we're
in USB_EP0_DATA_IN state and there is data to send.